[BUGFIX] Warn if exception dump could not be written
The System Logger now logs a warning if an exception
backtrace could not be written into Logs/Exceptions/
because the directory could not be created or is not
writeable.

Previously the attempt to write the exception backtrace
resulted in an exception. Uncool.

Original-Commit-Hash: 81ca141388c6b97c9d50cc01989d55de2bb1fe43

+++ TYPO3.Flow/Classes/Log/Logger.php
@@ -141,11 +141,13 @@ class Logger implements \TYPO3\FLOW3\Log\SystemLoggerInterface, \TYPO3\FLOW3\Log
 		if (!file_exists(FLOW3_PATH_DATA . 'Logs/Exceptions')) {
 			mkdir(FLOW3_PATH_DATA . 'Logs/Exceptions');
 		}
-		if (file_exists(FLOW3_PATH_DATA . 'Logs/Exceptions')) {
+		if (file_exists(FLOW3_PATH_DATA . 'Logs/Exceptions') && is_dir(FLOW3_PATH_DATA . 'Logs/Exceptions') && is_writable(FLOW3_PATH_DATA . 'Logs/Exceptions')) {
 			$referenceCode = ($exception instanceof \TYPO3\FLOW3\Exception) ? $exception->getReferenceCode() : $_SERVER['REQUEST_TIME'] . substr(md5(rand()), 0, 6);
 			$exceptionDumpPathAndFilename = FLOW3_PATH_DATA . 'Logs/Exceptions/' . $referenceCode . '.txt';
 			file_put_contents($exceptionDumpPathAndFilename, \TYPO3\FLOW3\var_dump($backTrace, 'Backtrace', TRUE, TRUE));
 			$message .= ' - See also: ' . basename($exceptionDumpPathAndFilename);
+		} else {
+			$this->log(sprintf('Could not write exception backtrace into %s because the directory could not be created or is not writable.', FLOW3_PATH_DATA . 'Logs/Exceptions/'), LOG_WARNING, array(), 'FLOW3', __CLASS__, __FUNCTION__);
 		}
 
 		$this->log($message, LOG_CRIT, $additionalData, $packageKey, $className, $methodName);
